Refactor:move Login,RegisterRequest to package dto
