+++ API/CapturedStyles/src/main/java/co/edu/ue/Auth/IAuthService.java
@@ -1,5 +1,8 @@
 package co.edu.ue.Auth;
 
+import co.edu.ue.dto.LoginRequest;
+import co.edu.ue.dto.RegisterRequest;
+
 public interface IAuthService {
 
   public AuthResponse login(LoginRequest request);
