+++ API/CapturedStyles/src/main/java/co/edu/ue/dto/RegisterRequest.java
@@ -1,4 +1,4 @@
-package co.edu.ue.Auth;
+package co.edu.ue.dto;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
